python/requirements.txt
requests==2.25.0
imparaai-checkers==1.4.2